Finally fixing *the* zoom bug!!
Previously, if you accidentally flicked at the end of a zoom (because you
lifted one finger slightly after the other and moved it before lifting), or
you switched fingers and moved it slightly, the image would sometimes fly or
reset to one corner or the other. No longer!

Conversely, if you panned after a zoom and attempted a flick on purpose,
there would be no momentum applied, because we had short-circuitted that
logic as an attempt to fix the above bug.

Both these things are fixed now! The fix was to properly reset the "start"
state on zoom end. This state includes the start finger, position and time.
Now momentum can be applied to flicks, and it also doesn't get incorrectly
applied on accidental flicks, etc.