Experimental: fix/tweak for improved mouse drag.
Allows dragging past the edge of the viewport now, and prevent some random
funkiness when moving the mouse quickly after a drag or click.

Seems to only make the mouse experience better, and doesn't seem to affect the
touch experience (touch events are captured natively already), but not sure if
I'm breaking something in the process...

The carousel example is a good test for this. Before this patch: dragging past
the edge of the carousel makes iScroll lose track of the mouse. After this
patch: iScroll continues to track the mouse.