Added a check if the scroller got stuck. Fixes #297
On Android when scrolling below the bottom line scroller gets stuck when the finger leaves the app at the top (the black bar).
This commit fixes this by peridocially (default 300ms) checking if the finger moved and if the scroller is beyond the bounds. If both are true, the scroller transitions from the stuck state to a state where it's scrolled down to the very bottom. This commit affects other devices and desktop browsers as well. But it makes sense to reset the scroller anyway if it's push beyond the bounds and doesn't move.