Fixed issue in Android with ending interaction failing.
Error occurred when passing over utils.preventDefaultException using the this.options.preventDefaultException.

Not only did the android find the ‘tagName : /^(INPUT…ECT)$/‘  as it should, but it found a function within this object which seemed to stump it.

Adding in the !utils.isBadAndroid (like in the _start method) stops it from happening.