Fixed #653.
The original fix provided for #653 simply disabled the functionally to fade out the indicator if on Android. My fix drops the use of Function.prototype.bind in favor of a local variable. I noticed that the variable 'val' was being passed into the function via bind. I'm not sure if there is any way that the variable 'val' can change after the setTimeout call is made, but if so, the fix would be to wrap the setTimeout call into another function passing in the 'val' variable as a parameter.