Add a forceDisplay option for indicators
If you are working with indicators for parallax effect on a responsive website (especially with horizontal navigation) you might want to prevent the indicators to be hidden when the container is not big enough to scroll. In this case, I've added the 'forceDisplay' option for the indicator. Just use 'forceDisplay: true' in the indicator option array.