Update _animate.js
I'm seeing a bug in Safari 12.0.2/OSX 10.13.6 in which this condition is not met, and therefore scrollEnd never gets fired. It only happens when the user gets to the end of the scroll pane, goes up a couple pages, then goes back to the end. If I remove the if statement, scrollEnd is always fired and the issue is resolved. Is there a hidden negative to doing that which I've not yet seen?